SystemMonitor: Use u64 for all GraphWidget values
Turns out size_t is not guaranteed to be 64-bit on i686 and trying to
set the max value using a u64 caused a narrowing conversion.

+++ Userland/Applications/SystemMonitor/GraphWidget.cpp
@@ -20,7 +20,7 @@ GraphWidget::~GraphWidget()
 {
 }
 
-void GraphWidget::add_value(Vector<size_t, 1>&& value)
+void GraphWidget::add_value(Vector<u64, 1>&& value)
 {
     m_values.enqueue(move(value));
     update();
